Move half/fifth Jinja filters.
It was possible to hit cases that did not use CoC7e but still expected
the filters to be available.
